--- cv_next/app/about/page.tsx
@@ -1,12 +1,0 @@
-import { Suspense } from "react";
-import { AboutLayout } from "./components/aboutLayout";
-
-export default function Page() {
-  return (
-    <main>
-      <Suspense>
-        <AboutLayout />
-      </Suspense>
-    </main>
-  );
-}
