+++ cv_next/components/layout/navbar/navbar.tsx
@@ -10,7 +10,7 @@ import { PopupToggle } from "./popupToggle";
 const navLinks = [
   {
     route: "Home",
-    path: "/about",
+    path: "/login",
   },
   {
     route: "Feed",
